requirements_ocr_ai.txt
pytesseract==0.3.10
Pillow==10.1.0
openai==1.3.0
python-dotenv==1.0.0
tqdm==4.66.1